Fix for Screen Blank issue when accepting call.
This fixes the issue of screen going blank while the user
is in the process of accepting a call, caused by 6d6523c51.
Adding the flag ON_AFTER_RELEASE to reset user activity timer
when wake lock is released. That will keep the screen on for
a while before it goes to blank.

Change-Id: I6aef162d94bb557c045a268ec23760d887c52859